+++ app/src/main/java/sasd97/java_blog/xyz/gaberlunzie/presentation/converter/ConverterPresenter.java
@@ -1,16 +1,8 @@
 package sasd97.java_blog.xyz.gaberlunzie.presentation.converter;
 
-import android.util.Log;
-
 import com.arellomobile.mvp.InjectViewState;
 import com.arellomobile.mvp.MvpPresenter;
-
-import sasd97.java_blog.xyz.gaberlunzie.navigation.Router;
-import sasd97.java_blog.xyz.gaberlunzie.navigation.fragments.FragmentCommand;
-import sasd97.java_blog.xyz.gaberlunzie.navigation.fragments.FragmentCommandDecorator;
-import sasd97.java_blog.xyz.gaberlunzie.navigation.fragments.Replace;
-import sasd97.java_blog.xyz.gaberlunzie.presentation.setup.SetupFragment;
-import sasd97.java_blog.xyz.gaberlunzie.utils.SwipeDetector;
+import com.arellomobile.mvp.presenter.InjectPresenter;
 
 /**
  * Created by alexander on 11/07/2017.
@@ -18,15 +10,4 @@ import sasd97.java_blog.xyz.gaberlunzie.utils.SwipeDetector;
 
 @InjectViewState
 public class ConverterPresenter extends MvpPresenter<ConverterView> {
-
-    private Router<FragmentCommand> router;
-
-    public void setRouter(Router<FragmentCommand> router) {
-        this.router = router;
-    }
-
-    public void openSetup() {
-        FragmentCommandDecorator fragmentCommand = new Replace(SetupFragment.getInstance());
-        router.pushForward(fragmentCommand);
-    }
 }
