+++ app/src/main/java/sasd97/java_blog/xyz/gaberlunzie/presentation/main/MainView.java
@@ -1,0 +1,10 @@
+package sasd97.java_blog.xyz.gaberlunzie.presentation.main;
+
+import com.arellomobile.mvp.MvpView;
+
+/**
+ * Created by alexander on 11/07/2017.
+ */
+
+public interface MainView extends MvpView {
+}
