+++ app/src/main/res/layout/activity_convert.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="sasd97.java_blog.xyz.gaberlunzie.presentation.converter.ConverterActivity">
+    tools:context="sasd97.java_blog.xyz.gaberlunzie.presentation.main.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
